# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateLocationRequest:
    properties:
      lat:
        example: -6.2
        type: number
      lon:
        example: 106.816666
        type: number
      speed:
        example: 45.5
        type: number
      ts:
        example: "2026-04-19T08:30:00Z"
        type: string
    required:
    - lat
    - lon
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@jalurberlian.id
    name: API Support
    url: http://www.jalurberlian.id/support
  description: Real-time fleet management and order tracking system for PT. Jalur
    Berlian Makassar
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Jalur Berlian Fleet Management API
  version: "1.0"
paths:
...
  /trips/{id}/location:
    get:
      consumes:
      - application/json
      description: Retrieve the most recent GPS location for a trip. Returns the latest
        coordinate point recorded.
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Latest location record with lat, lon, speed, timestamp
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid trip ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found - no location recorded for this trip
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - failed to retrieve latest location
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get latest location
      tags:
      - Locations
    post:
      consumes:
      - application/json
      description: Save current GPS location for a trip. Accepts latitude, longitude,
        optional speed, and timestamp. Used for real-time vehicle tracking.
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Location data: lat (float64 required), lon (float64 required),
          speed (float64 optional), ts (RFC3339 string optional, defaults to now)'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Location saved successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request - invalid trip ID or invalid lat/lon coordinates
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Save trip location
      tags:
      - Locations
  /trips/{id}/locations:
    get:
      consumes:
      - application/json
      description: Retrieve the location history for a trip. Returns all saved GPS
        coordinates in descending order (latest first). Supports pagination via limit
        parameter.
      parameters:
      - description: Trip ID
        in: path
        name: id
        required: true
        type: integer
      - description: Max number of locations to return (default 50, max 500)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Array of location records with lat, lon, speed, timestamp
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid trip ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error - failed to retrieve location history
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get location history
      tags:
      - Locations
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"